+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The following is a curated list of changes in the Enact agate module, newest cha
 ### Added
 
 - `agate/Heading` prop `showBackButton`
-- `agate/Input` prop `clearInputButton` and `clearInputIcon`
+- `agate/Input` prop `clearButton` and `clearIcon`
 - `agate/Keypad` prop `activeCall` to render different icons depending on whether there is an active call or not
 
 ### Fixed